% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-wrappers.R
\name{bs_col}
\alias{bs_col}
\title{Construct bootstrap column classes}
\usage{
bs_col(
  base = 12,
  small = NULL,
  medium = NULL,
  large = NULL,
  extra_large = NULL,
  more = NULL
)
}
\arguments{
\item{base}{Default number of columns}

\item{small}{Number of columns on small devices}

\item{medium}{Number of columns on medium devices}

\item{large}{Number of columns on large devices}

\item{extra_large}{Number of columns on extra_large devices}

\item{more}{Additional classes to add}
}
\value{

}
\description{
Construct bootstrap column classes
}
