% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-wrappers.R
\name{bs_mar}
\alias{bs_mar}
\title{Construct bootstrap margin classes}
\usage{
bs_mar(base = 0, bottom = NULL, top = NULL, left = NULL, right = NULL)
}
\arguments{
\item{base}{Margin size}

\item{bottom}{Margin size on bottom}

\item{top}{Margin size on top}

\item{left}{Margin size on left}

\item{right}{Margin size on right}
}
\value{

}
\description{
Construct bootstrap margin classes
}
