% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bs_pad}
\alias{bs_pad}
\title{Construct bootstrap padding classes}
\usage{
bs_pad(base = 0, bottom = NULL, top = NULL, left = NULL, right = NULL)
}
\arguments{
\item{base}{Padding size}

\item{bottom}{Padding size on bottom}

\item{top}{Padding size on top}

\item{left}{Padding size on left}

\item{right}{Padding size on right}
}
\value{

}
\description{
Construct bootstrap padding classes
}
