% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-builder.R
\name{card_columns}
\alias{card_columns}
\title{Create columns of "masonry-like" cards}
\usage{
card_columns(cards)
}
\arguments{
\item{cards}{List of cards}
}
\value{
A thing
}
\description{
Create columns of "masonry-like" cards
}
