% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-parts.R
\name{card_footer}
\alias{card_footer}
\title{Specify the footer of a card}
\usage{
card_footer(text)
}
\arguments{
\item{text}{String specifying footer content}
}
\value{
A "shiny.tag" object
}
\description{
Specify the footer of a card
}
