% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-builder.R
\name{card_group}
\alias{card_group}
\title{Create a group of cards}
\usage{
card_group(cards, margin = bs_mar(top = 2, bottom = 2), padding = bs_pad())
}
\arguments{
\item{cards}{List of cards}

\item{margin}{Bootstrap classes setting card group margin}

\item{padding}{Bootstrap classes setting card group padding}
}
\value{
A thing
}
\description{
Create a group of cards
}
