% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-parts.R
\name{card_header}
\alias{card_header}
\title{Specify the header of a card}
\usage{
card_header(text)
}
\arguments{
\item{text}{String specifying header content}
}
\value{
A "shiny.tag" object
}
\description{
Specify the header of a card
}
