% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-parts.R
\name{card_subtitle}
\alias{card_subtitle}
\title{Specify the subtitle of a card}
\usage{
card_subtitle(subtitle, level = 6)
}
\arguments{
\item{subtitle}{String specifying the subtitle}

\item{level}{Level of header (integer between 1 and 6, default = 6)}
}
\value{
A "shiny.tag" object
}
\description{
Specify the subtitle of a card
}
