% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-parts.R
\name{card_text}
\alias{card_text}
\title{Specify the text of a card}
\usage{
card_text(text)
}
\arguments{
\item{title}{String specifying the text to insert in a card}
}
\value{
A "shiny.tag" object
}
\description{
Specify the text of a card
}
