% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-parts.R
\name{card_title}
\alias{card_title}
\title{Specify the title of a card}
\usage{
card_title(title, level = 5)
}
\arguments{
\item{title}{String specifying the title}

\item{level}{Level of header (integer between 1 and 6, default = 5)}
}
\value{
A "shiny.tag" object
}
\description{
Specify the title of a card
}
