% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cards}
\alias{cards}
\title{Specifies a bootstrap card}
\usage{
cards(
  data,
  title = NULL,
  text = NULL,
  image = NULL,
  link = NULL,
  footer = NULL,
  header = NULL,
  width = "medium",
  label_position = "below",
  padding = 0,
  margin = 1
)
}
\arguments{
\item{data}{Data}

\item{title}{Title for the card}

\item{text}{Text for the card}

\item{image}{Path to the card image}

\item{link}{URL to link to from title and image}

\item{footer}{Card footer}

\item{header}{Card header}

\item{width}{Card width ("narrow", "medium", "wide")}

\item{label_position}{Position of the label relative to the image (e.g., "below", "above", "left", "right")}

\item{padding}{Spacing between parts of the card (integer between 0 and 5)}

\item{margin}{Spacing between adjacent cards (integer between 0 and 5)}
}
\value{
A "shiny.tag" object
}
\description{
Specifies a bootstrap card
}
