% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galleries.R
\docType{data}
\name{galleries}
\alias{galleries}
\title{Generative art galleries}
\format{
A tibble with 7 rows and 5 variables:
\describe{
\item{long_name}{full name of the gallery when released}
\item{short_name}{compact name used to link to gallery and images}
\item{date}{date on which the gallery was posted}
\item{blurb}{brief description of the gallery}
\item{paintbrushes}{names of R packages used to generate the art}
}
}
\source{
\url{https://art.djnavarro.net/}
}
\usage{
galleries
}
\description{
A dataset containing information describing several galleries of generative
art created in R.
}
\keyword{datasets}
