% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sample_card}
\alias{sample_card}
\title{Generate a random kitten card}
\usage{
sample_card(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{card()}}
}
\value{
Specification for a bootstrap card
}
\description{
Generate a random kitten card
}
