% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sample_lorem}
\alias{sample_lorem}
\title{Generate lorem ipsum text}
\usage{
sample_lorem(use_cache = TRUE)
}
\arguments{
\item{use_cache}{Sample from cached lorem ipusm text? (default = TRUE)}
}
\value{
A string containing the text
}
\description{
Generate lorem ipsum text
}
