% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strapless.R
\name{use_strapless}
\alias{use_strapless}
\title{Use the strapless css library}
\usage{
use_strapless(destdir = ".", source = "local")
}
\arguments{
\item{destdir}{directory to install the strapless library}

\item{source}{set source = "local" to use the cached copy, source = "github" to download}
}
\value{
path to file
}
\description{
Use the strapless css library
}
