% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-builder.R
\name{card}
\alias{card}
\title{Specifies a bootstrap card}
\usage{
card(title, image, text)
}
\arguments{
\item{title}{Title for the card}

\item{image}{Path to the card image}

\item{text}{Text for the card}
}
\value{
A "shiny.tag" object
}
\description{
Specifies a bootstrap card
}
