% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-builder.R
\name{card}
\alias{card}
\title{Specifies a bootstrap card}
\usage{
card(title, text, image, image_align = "top", col_spec = "col-12 col-md-4")
}
\arguments{
\item{title}{Title for the card}

\item{text}{Text for the card}

\item{image}{Path to the card image}

\item{image_align}{Placement of the card image ("top" or "bottom")}

\item{col_spec}{Bootstrap classes setting card width}
}
\value{
A "shiny.tag" object
}
\description{
Specifies a bootstrap card
}
