% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-builder.R
\name{card}
\alias{card}
\title{Specifies a bootstrap card}
\usage{
card(
  title,
  text,
  image,
  image_align = "top",
  width = bs_col(medium = 4),
  margin = bs_mar(base = 3),
  padding = bs_pad(),
  footer = NULL,
  header = NULL
)
}
\arguments{
\item{title}{Title for the card}

\item{text}{Text for the card}

\item{image}{Path to the card image}

\item{image_align}{Placement of the card image ("top" or "bottom")}

\item{width}{Bootstrap classes setting card width}

\item{margin}{Bootstrap classes setting card margin}

\item{padding}{Bootstrap classes setting card padding}

\item{footer}{Card footer}

\item{header}{Card header}
}
\value{
A "shiny.tag" object
}
\description{
Specifies a bootstrap card
}
