% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-builder.R
\name{card}
\alias{card}
\title{Specifies a bootstrap card}
\usage{
card(title, image, text, width = 25)
}
\arguments{
\item{title}{Title for the card}

\item{image}{Path to the card image}

\item{text}{Text for the card}

\item{width}{Width for the card (25, 50, 75 or 100, default = 25)}
}
\value{
A "shiny.tag" object
}
\description{
Specifies a bootstrap card
}
