% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-builder.R
\name{card_grid}
\alias{card_grid}
\title{Create a grid of cards}
\usage{
card_grid(cards, margin = bs_mar(top = 3, bottom = 3), padding = bs_pad())
}
\arguments{
\item{cards}{List of cards}

\item{margin}{Bootstrap classes setting card group margin}

\item{padding}{Bootstrap classes setting card group padding}
}
\value{
A thing
}
\description{
Create a grid of cards
}
