% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-arranger.R
\name{card_grid}
\alias{card_grid}
\title{Create a deck of cards}
\usage{
card_grid(cards, type = "card-group")
}
\arguments{
\item{cards}{List of cards}

\item{type}{Bootstrap class to define the grid ("card-group" or "card-deck")}
}
\value{
sdfgdfg
}
\description{
Create a deck of cards
}
