% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-parts.R
\name{card_image}
\alias{card_image}
\title{Specify the image for a card}
\usage{
card_image(
  image,
  link = NULL,
  alt = NULL,
  align = "top",
  padding = 0,
  margin = 0
)
}
\arguments{
\item{image}{Path to the image file}

\item{link}{URL to visit when image is clicked (optional)}

\item{alt}{Alt text for the image}

\item{align}{Alignment of the image within the card ("top", "left" or "right")}

\item{padding}{Padding size for the image (integer between 0 and 5)}

\item{margin}{Margin size for the image (integer between 0 and 5)}
}
\value{
A "shiny.tag" object
}
\description{
Specify the image for a card
}
