% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cards}
\alias{cards}
\title{Specifies a bootstrap card}
\usage{
cards(
  data,
  title = NULL,
  text = NULL,
  image = NULL,
  link = NULL,
  footer = NULL,
  header = NULL,
  width = "medium",
  overlay = "FALSE"
)
}
\arguments{
\item{data}{Data}

\item{title}{Title for the card}

\item{text}{Text for the card}

\item{image}{Path to the card image}

\item{link}{URL to link to from title and image}

\item{footer}{Card footer}

\item{header}{Card header}

\item{width}{Card width ("narrow", "medium", "wide")}

\item{overlay}{Logical}
}
\value{
A "shiny.tag" object
}
\description{
Specifies a bootstrap card
}
