% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cards}
\alias{cards}
\title{Specifies a bootstrap card}
\usage{
cards(
  data,
  title = NULL,
  text = NULL,
  image = NULL,
  link = NULL,
  footer = NULL,
  tags = NULL,
  width = "medium",
  layout = "label-below",
  padding = 0,
  gutter = 1,
  breakpoint = NULL,
  colour = "#ffffffaa",
  border_width = "1px",
  border_colour = "#808080",
  rounding = "1rem"
)
}
\arguments{
\item{data}{Data}

\item{title}{Title for the card}

\item{text}{Text for the card}

\item{image}{Path to the card image}

\item{link}{URL to link to from title and image}

\item{footer}{Card footer}

\item{tags}{Tags to be assigned to each card}

\item{width}{Card width ("narrow", "medium", "wide")}

\item{layout}{Card layout ("label-above", "label-below", "label-left", "label-right", "inset-top", "inset-bottom")}

\item{padding}{Spacing between parts of the card (integer between 0 and 5)}

\item{gutter}{Spacing between adjacent cards (integer between 0 and 5)}

\item{breakpoint}{Position to insert breaks for horizontal or inset cards, interpreted as proportion of card allocated to label}

\item{colour}{Colour applied to the card (interpretation depends on layout)}

\item{border_width}{Width of card border}

\item{border_colour}{Colour of card border}

\item{rounding}{Amount of rounding on card corners}
}
\value{
A "shiny.tag" object
}
\description{
Specifies a bootstrap card
}
