% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cards}
\alias{cards}
\title{Specifies a bootstrap card}
\usage{
cards(
  data,
  title = NULL,
  text = NULL,
  image = NULL,
  link = NULL,
  footer = NULL,
  header = NULL,
  tags = NULL,
  layout = "label-below",
  width = 3,
  spacing = 2,
  breakpoint = 4,
  label_colour = "#ffffffaa",
  border_colour = "#808080",
  border_width = 1,
  rounding = 3
)
}
\arguments{
\item{data}{Data}

\item{title}{Title for the card}

\item{text}{Text for the card}

\item{image}{Path to the card image}

\item{link}{URL to link to from title and image}

\item{footer}{Card footer}

\item{header}{Card header}

\item{tags}{Tags to be assigned to each card}

\item{layout}{Card layout is a string, defaults to "label-below" (see details)}

\item{width}{Card width is an integer between 1 and 5}

\item{spacing}{Spacing between cards is an integer between 0 and 5}

\item{breakpoint}{Number between 1 and 5, controlling label size on horizontal and inset layouts}

\item{label_colour}{Colour applied to the card label}

\item{border_colour}{Colour applied to the card border}

\item{border_width}{Width of card border is an integer between 0 and 5}

\item{rounding}{Amount of rounding on card corners is an integer between 0 and 5}
}
\value{
A "shiny.tag" object
}
\description{
Specifies a bootstrap card
}
