% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cards}
\alias{cards}
\title{Builds a deck of bootstrap cards}
\usage{
cards(
  data,
  title = NULL,
  text = NULL,
  image = NULL,
  link = NULL,
  footer = NULL,
  header = NULL,
  tags = NULL,
  layout = "label-below",
  width = 3,
  spacing = 2,
  breakpoint = 4,
  label_colour = NULL,
  border_colour = NULL,
  border_width = 1,
  border_radius = 3
)
}
\arguments{
\item{data}{Data frame}

\item{title}{Title for the card}

\item{text}{Text for the card}

\item{image}{Path to the card image}

\item{link}{URL to link to from title and image}

\item{footer}{Card footer}

\item{header}{Card header}

\item{tags}{Tags to be assigned to each card}

\item{layout}{Card layout is a string, defaults to "label-below" (see details)}

\item{width}{Card width is an integer between 1 and 5}

\item{spacing}{Spacing between cards is an integer between 0 and 5}

\item{breakpoint}{Number between 1 and 5, controlling label size on horizontal and inset layouts}

\item{label_colour}{Colour applied to the card label (if NULL, colour inherits)}

\item{border_colour}{Colour applied to the card border (if NULL, colour inherits)}

\item{border_width}{Width of card border is an integer between 0 and 5}

\item{border_radius}{Amount of rounding on card corners is an integer between 0 and 5}
}
\value{
A "shiny.tag" object containing the HTML for the card deck
}
\description{
Builds a deck of bootstrap cards
}
\details{
This function constructs the HTML necessary to specify a deck of bootstrap 4
cards suitable for inclusion within an R markdown document. It takes a data
frame as the first argument, with one row for each card to be generated.

Cards are specified using the \code{title}, \code{text}, \code{image},
\code{link}, \code{footer}, \code{header}, and \code{tags} arguments. These
arguments take expressions to be evaluated using the user-supplied
\code{data}, and should evaluate to character vectors that have length 1 or
the same number of rows as \code{data}.

The \code{layout} argument is a single character string specifying the
layout of the cards: possible values are "label-below" (the default),
"label-above", "label-left", "label-right", "label-only", "image-only",
"inset-top", "inset-bottom".

The \code{width}, \code{spacing}, \code{breakpoint}, \code{border_width},
and \code{border_radius} arguments all take integer inputs specifying
the visual appearance of the cards. Allowed values range from 0 to 5, except
for \code{width} and \code{breakpoint} which cannot be 0. The other two
arguments \code{label_colour} and \code{border_colour} take a single string
specifying colours of the relevant parts of the card.
}
\examples{
\dontrun{
galleries \%>\%
  cards(
    title = long_name,
    text = blurb,
    image = image_url,
    link = gallery_url
  )
}
}
